Fix detection of preferred MIME type
Previously, the algorithm would not take into account the order of
appearance of matching content types in the `Accept` header.

There was a test describing the expected behavior. It was green
because the order of the content types in the example happened to
match the order in which the supported formats were checked.